# savicar-api/docs/swagger.yaml
basePath: /
definitions:
...
  contact.Contact:
    properties:
      address:
        type: string
      address_number:
        type: integer
      email:
        type: string
      id_city:
        type: integer
      id_contact:
        type: integer
      id_customer:
        type: integer
      is_mobile_phone_whatsapp:
        type: boolean
      mobile_phone:
        type: string
      neighborhood:
        type: string
    type: object
  customer.Customer:
    properties:
      dob:
        type: string
      gender:
        type: integer
      id_customer:
        type: integer
      image_path:
        type: string
      individual_name:
        type: string
      is_active:
        type: boolean
      is_legal_person:
        type: boolean
      legal_name:
        type: string
      municipal_registration:
        type: string
      state_registration:
        type: string
      tax_id:
        type: string
      trade_name:
        type: string
      web_site:
        type: string
    type: object
  customermodel.CustomerModel:
    properties:
      color:
        type: string
      customer_name:
        type: string
      id_customer:
        type: integer
      id_customer_model:
        type: integer
      id_model:
        type: integer
      model_name:
        type: string
      plate:
        type: string
      vin:
        type: string
      year_make:
        type: integer
      year_model:
        type: integer
    type: object
...
  vehiclemodel.VehicleModel:
    properties:
      id_fuel:
        type: integer
      id_make:
        type: integer
      id_model:
        type: integer
      name:
        type: string
      version:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Service Order Images API
  title: Badencar API
  version: "1.0"
paths:
  /contacts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/contact.Contact'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contact.Contact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/contact.Contact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a contact
      tags:
      - contacts
    get:
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contact.Contact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a contact by ID
      tags:
      - contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contact.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contact.Contact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a contact
      tags:
      - contacts
  /customer-models:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customermodel.CustomerModel'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all customer models
      tags:
      - customer-models
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customermodel.CustomerModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customermodel.CustomerModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a customer model
      tags:
      - customer-models
  /customer-models/{id}:
    delete:
      parameters:
      - description: Customer Model ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a customer model
      tags:
      - customer-models
    get:
      parameters:
      - description: Customer Model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customermodel.CustomerModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a customer model by ID
      tags:
      - customer-models
    put:
      consumes:
      - application/json
      parameters:
      - description: Customer Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customermodel.CustomerModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customermodel.CustomerModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a customer model
      tags:
      - customer-models
  /customers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customer.Customer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customer.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.Customer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a customer
      tags:
      - customers
  /customers/{id}:
    delete:
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a customer
      tags:
      - customers
    get:
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.Customer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a customer by ID
      tags:
      - customers
    put:
      consumes:
      - application/json
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customer.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.Customer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a customer
      tags:
      - customers
...
  /models:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vehiclemodel.VehicleModel'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all vehicle models
      tags:
      - models
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/vehiclemodel.VehicleModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/vehiclemodel.VehicleModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a vehicle model
      tags:
      - models
  /models/{id}:
    delete:
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a vehicle model
      tags:
      - models
    get:
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vehiclemodel.VehicleModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a vehicle model by ID
      tags:
      - models
    put:
      consumes:
      - application/json
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/vehiclemodel.VehicleModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vehiclemodel.VehicleModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a vehicle model
      tags:
      - models
...
swagger: "2.0"